diting: parts: Implement AOD auto brightness service
AOSP's doze auto brightness mechanism doesn't work for us since our
HWC does not allow updating brightness when the display is not on.
Implement a sensor listener ourselves and set doze to high or low
brightness (HBM or LBM) depending on the sensor reading.

While we're at it, workaround a bug that causes display to turn off
randomly while in AOD by setting display driver's doze_brightness to
-1 (invalid).

Signed-off-by: klozz <carlosj@klozz.dev>