diting: Drop qvr/qxr
Does not seem that useful, and conflicts with the
libgrpc++_unsecure package if elf checks are turned
on in extract files.
The libgrpc++_unsecure package builds into the
libgrpc++_unsecure-1.61.1.so binary, but the QVR blob
needs libgrpc++_unsecure.so.
Rather than mess with extract files to fix it
some more, just get rid of it.

Change-Id: I53556a5b3452138341de6e1c1556aabb7cd8d711

Signed-off-by: klozz <carlosj@klozz.dev>