diting: udfps: Disable lhbm when authentication is  finished
We previously missed the case when the finger is rejected the
fifth time and neither cancel nor a successful acquired message
was sent, leading to lhbm staying on forever. These new methods
should cover all cases in which the authentication is over. We
do not need to handle onAcquired and cancel explicitly anymore,
so drop the corresponding methods.

Change-Id: Ibc284bc250e60a46d220f6a8f4c82908cd1f0776

Change-Id: I8a83ff4cc2f29dd37e665dcde5fa19b60ec35cbe
Signed-off-by: klozz <carlosj@klozz.dev>